pyproject.toml
[project]
name = "telegram-bot-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.24.0",
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "greenlet>=3.3.0",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.46",
    "var-dump>=1.2",
    "yookassa>=3.10.0",
]

[tool.pyright]
typeCheckingMode = "standard"
